+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next
+
+### Changed
+
+* **Deprecated** `logging.Handler`, you should use `logging.NewAddTraceIDMiddleware` instead.
+
 ## 2020-03-21
 
 ### Changed